Refactors env modules
