--- src/rlplg/environments/alphabet/constants.py
@@ -1,4 +1,0 @@
-ENV_NAME = "ABCSeq"
-MIN_SEQ_LENGTH = 1
-LETTERS = [chr(value) for value in range(ord("A"), ord("Z") + 1)]
-NUM_LETTERS = len(LETTERS)
