--- src/rlplg/environments/redgreen/constants.py
@@ -1,9 +1,0 @@
-ENV_NAME = "RedGreenSeq"
-RED_PILL = 0
-GREEN_PILL = 1
-WAIT = 2
-ACTIONS = [RED_PILL, GREEN_PILL, WAIT]
-
-ACTION_NAME_MAPPING = {"red": RED_PILL, "green": GREEN_PILL, "wait": WAIT}
-OBS_KEY_CURE_SEQUENCE = "cure_sequence"
-OBS_KEY_POSITION = "position"
