+++ tests/rlplg/environments/test_abcseq.py
@@ -5,12 +5,12 @@ import pytest
 from tf_agents.specs import array_spec
 from tf_agents.trajectories import time_step as ts
 
-from rlplg.environments.alphabet import env
+from rlplg.environments import abcseq
 
 
 @hypothesis.given(length=st.integers(min_value=1, max_value=100))
 def test_abcseq_init(length: int):
-    environment = env.ABCSeq(length)
+    environment = abcseq.ABCSeq(length)
     assert environment.length == length
     assert environment.action_spec() == action_spec(length)
     assert environment.observation_spec() == observation_spec(length)
@@ -18,7 +18,7 @@ def test_abcseq_init(length: int):
 
 def test_abcseq_simple_sequence():
     length = 4
-    environment = env.ABCSeq(length)
+    environment = abcseq.ABCSeq(length)
     assert_time_step(
         environment.reset(),
         ts.TimeStep(
@@ -124,7 +124,7 @@ def test_abcseq_simple_sequence():
     length=st.integers(min_value=1, max_value=100),
 )
 def test_abcseq_render(length: int):
-    environment = env.ABCSeq(length)
+    environment = abcseq.ABCSeq(length)
     environment.reset(),
     # starting point
     np.testing.assert_array_equal(
@@ -142,7 +142,7 @@ def test_abcseq_render(length: int):
 )
 def test_abcseq_render_with_invalid_modes(length: int):
     modes = ("human",)
-    environment = env.ABCSeq(length)
+    environment = abcseq.ABCSeq(length)
     for mode in modes:
         with pytest.raises(NotImplementedError):
             environment.render(mode)
@@ -158,7 +158,7 @@ def test_apply_action_with_unmastered_letters_and_action_is_the_next(
     obs = np.array([1] + [1] * completed + [0] * missing)
     expected = np.array([1] + [1] * (completed + 1) + [0] * (missing - 1))
     action = completed
-    output = env.apply_action(obs, action)
+    output = abcseq.apply_action(obs, action)
     np.testing.assert_array_equal(output, expected)
 
 
@@ -172,7 +172,7 @@ def test_apply_action_with_unmastered_letters_and_action_is_skipped_head(
     obs = np.array([1] + [1] * completed + [0] * missing)
     skipped_steps = np.random.randint(low=1, high=missing + 1)
     action = completed + skipped_steps
-    output = env.apply_action(obs, action)
+    output = abcseq.apply_action(obs, action)
     np.testing.assert_array_equal(output, obs)
 
 
@@ -186,7 +186,7 @@ def test_apply_action_with_unmastered_letters_and_action_is_behind(
     obs = np.array([1] + [1] * completed + [0] * missing)
     action = np.random.randint(low=0, high=completed)
     # num letters ahead + behind + 1 for rotation
-    output = env.apply_action(obs, action)
+    output = abcseq.apply_action(obs, action)
     np.testing.assert_array_equal(output, obs)
 
 
@@ -196,7 +196,7 @@ def test_apply_action_with_unmastered_letters_and_action_is_behind(
 def test_apply_action_with_completed_sequence_action_is_end(completed: int):
     obs = np.array([1] + [1] * completed)
     action = completed
-    output = env.apply_action(obs, action)
+    output = abcseq.apply_action(obs, action)
     np.testing.assert_array_equal(output, obs)
 
 
@@ -209,7 +209,7 @@ def test_action_reward_with_unmastered_letters_and_action_is_the_next(
 ):
     obs = np.array([1] + [1] * completed + [0] * missing)
     action = completed
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     assert output == -1.0
 
 
@@ -221,7 +221,7 @@ def test_action_reward_with_one_unmastered_letters_and_action_is_the_next(
 ):
     obs = np.array([1] + [1] * completed + [0])
     action = completed
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     assert output == -1.0
 
 
@@ -235,7 +235,7 @@ def test_action_reward_with_unmastered_letters_and_action_is_skipped_ahead(
     obs = np.array([1] + [1] * completed + [0] * missing)
     skipped_steps = np.random.randint(low=1, high=missing + 1)
     action = completed + skipped_steps
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     moving_penantly = 1
     assert output == -(skipped_steps + moving_penantly)
 
@@ -252,7 +252,7 @@ def test_action_reward_with_unmastered_letters_and_action_is_behind(
     # num letters ahead + behind + 1 for rotation
     moving_penalty = 1
     expected = missing + action + moving_penalty + 1
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     assert output == -expected
 
 
@@ -262,7 +262,7 @@ def test_action_reward_with_unmastered_letters_and_action_is_behind(
 def test_action_reward_with_completed_sequence_action_is_end(completed: int):
     obs = np.array([1] + [1] * completed)
     action = completed
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     assert output == 0.0
 
 
@@ -272,7 +272,7 @@ def test_action_reward_with_completed_sequence_action_is_end(completed: int):
 def test_action_reward_with_completed_sequence_action_is_behind(completed: int):
     obs = np.array([1] + [1] * completed)
     action = np.random.randint(low=0, high=completed)
-    output = env.action_reward(obs, action)
+    output = abcseq.action_reward(obs, action)
     assert output == 0.0
 
 
@@ -281,24 +281,24 @@ def test_action_reward_with_completed_sequence_action_is_behind(completed: int):
 )
 def test_beginning_state(length: int):
     np.testing.assert_array_equal(
-        env.beginning_state(length), np.array([1] + [0] * length)
+        abcseq.beginning_state(length), np.array([1] + [0] * length)
     )
 
 
 def test_is_finished():
-    assert not env.is_finished(np.array([1, 0]), 0)
-    assert not env.is_finished(np.array([1, 0]), 1)
-    assert env.is_finished(np.array([1, 1]), 0)
-    assert not env.is_finished(np.array([1, 0, 0]), 0)
-    assert not env.is_finished(np.array([1, 1, 0]), 0)
-    assert env.is_finished(np.array([1, 1, 1]), 1)
+    assert not abcseq.is_finished(np.array([1, 0]), 0)
+    assert not abcseq.is_finished(np.array([1, 0]), 1)
+    assert abcseq.is_finished(np.array([1, 1]), 0)
+    assert not abcseq.is_finished(np.array([1, 0, 0]), 0)
+    assert not abcseq.is_finished(np.array([1, 1, 0]), 0)
+    assert abcseq.is_finished(np.array([1, 1, 1]), 1)
 
 
 def test_state_id():
     # (starting state)
-    assert env.get_state_id([1, 0, 0]) == 0
-    assert env.get_state_id([1, 1, 0]) == 1
-    assert env.get_state_id([1, 1, 1]) == 2
+    assert abcseq.get_state_id([1, 0, 0]) == 0
+    assert abcseq.get_state_id([1, 1, 0]) == 1
+    assert abcseq.get_state_id([1, 1, 1]) == 2
 
 
 def action_spec(length: int) -> array_spec.BoundedArraySpec:
